Complete stateful store cutover
