+++ crates/tandem-server/src/app/tasks.rs
@@ -16,7 +16,8 @@ use crate::runtime_event_log::{
     append_runtime_event_log_row, prune_runtime_event_log, RuntimeEventLogRow,
 };
 use crate::stateful_runtime::{
-    compact_stateful_run_event_log, prune_stateful_wait_store, StatefulRuntimeStoragePaths,
+    compact_stateful_run_event_log, prune_stateful_reliability_store, prune_stateful_wait_store,
+    stateful_reliability_path_from_runtime_events_path, StatefulRuntimeStoragePaths,
 };
 use crate::util::time::now_ms;
 
@@ -291,465 +292,13 @@ fn event_tenant_context_value(event: &EngineEvent) -> Value {
         .unwrap_or(Value::Null)
 }
 
-fn session_context_run_event_input(event: &EngineEvent) -> Option<ContextRunEventAppendInput> {
-    match event.event_type.as_str() {
-        "session.run.started" => Some(ContextRunEventAppendInput {
-            event_type: "session_run_started".to_string(),
-            status: ContextRunStatus::Running,
-            step_id: Some("session-run".to_string()),
-            payload: serde_json::json!({
-                "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                "runID": event.properties.get("runID").cloned().unwrap_or(Value::Null),
-                "agentID": event.properties.get("agentID").cloned().unwrap_or(Value::Null),
-                "agentProfile": event.properties.get("agentProfile").cloned().unwrap_or(Value::Null),
-                "tenantContext": event_tenant_context_value(event),
-                "why_next_step": "session run in progress",
-                "step_status": "in_progress",
-            }),
-        }),
-        "message.part.updated" => {
-            if is_running_tool_args_delta(&event.properties) {
-                return None;
-            }
-            let part = event.properties.get("part")?;
-            let part_type = part
-                .get("type")
-                .and_then(|value| value.as_str())
-                .unwrap_or_default();
-            if !matches!(
-                part_type,
-                "tool" | "tool-invocation" | "tool-result" | "tool_invocation" | "tool_result"
-            ) {
-                return None;
-            }
-            let tool_name = part
-                .get("tool")
-                .and_then(|value| value.as_str())
-                .unwrap_or("tool");
-            let tool_state = part
-                .get("state")
-                .and_then(|value| value.as_str())
-                .unwrap_or("running");
-            let why_next_step = match tool_state {
-                "completed" => format!("tool `{tool_name}` completed"),
-                "failed" => format!("tool `{tool_name}` failed"),
-                _ => format!("tool `{tool_name}` running"),
-            };
-            Some(ContextRunEventAppendInput {
-                event_type: "session_tool_updated".to_string(),
-                status: ContextRunStatus::Running,
-                step_id: Some("session-run".to_string()),
-                payload: serde_json::json!({
-                    "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                    "runID": event.properties.get("runID").cloned().unwrap_or(Value::Null),
-                    "part": compact_context_tool_value(Some(part)),
-                    "toolCallDelta": compact_context_tool_value(event.properties.get("toolCallDelta")),
-                    "tenantContext": event_tenant_context_value(event),
-                    "why_next_step": why_next_step,
-                    "step_status": if tool_state == "completed" { "done" } else { "in_progress" },
-                    "error": part.get("error").cloned().unwrap_or(Value::Null),
-                }),
-            })
-        }
-        "session.run.finished" => {
-            let status = event
-                .properties
-                .get("status")
-                .and_then(|value| value.as_str())
-                .unwrap_or("completed");
-            Some(ContextRunEventAppendInput {
-                event_type: "session_run_finished".to_string(),
-                status: session_run_status_to_context(status),
-                step_id: Some("session-run".to_string()),
-                payload: serde_json::json!({
-                    "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                    "runID": event.properties.get("runID").cloned().unwrap_or(Value::Null),
-                    "status": status,
-                    "error": event.properties.get("error").cloned().unwrap_or(Value::Null),
-                    "tenantContext": event_tenant_context_value(event),
-                    "why_next_step": format!("session run finished with status `{status}`"),
-                    "step_status": if matches!(status, "completed") { "done" } else if matches!(status, "cancelled") { "blocked" } else { "failed" },
-                }),
-            })
-        }
-        "tool.effect.recorded" => {
-            let record = event.properties.get("record")?;
-            let tool = record
-                .get("tool")
-                .and_then(|value| value.as_str())
-                .unwrap_or("tool");
-            let status = record
-                .get("status")
-                .and_then(|value| value.as_str())
-                .unwrap_or("started");
-            let phase = record
-                .get("phase")
-                .and_then(|value| value.as_str())
-                .unwrap_or("invocation");
-            let summary = match status {
-                "succeeded" => format!("tool `{tool}` {phase} succeeded"),
-                "failed" => format!("tool `{tool}` {phase} failed"),
-                "blocked" => format!("tool `{tool}` {phase} blocked"),
-                _ => format!("tool `{tool}` {phase} started"),
-            };
-            Some(ContextRunEventAppendInput {
-                event_type: "tool_effect_recorded".to_string(),
-                status: ContextRunStatus::Running,
-                step_id: Some("session-run".to_string()),
-                payload: serde_json::json!({
-                    "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                    "messageID": event.properties.get("messageID").cloned().unwrap_or(Value::Null),
-                    "tool": event.properties.get("tool").cloned().unwrap_or(Value::Null),
-                    "record": record.clone(),
-                    "tenantContext": event_tenant_context_value(event),
-                    "why_next_step": summary,
-                    "step_status": if matches!(status, "failed" | "blocked") {
-                        "blocked"
-                    } else {
-                        "in_progress"
-                    },
-                }),
-            })
-        }
-        "tool.routing.decision" => Some(ContextRunEventAppendInput {
-            event_type: "tool_routing_decision".to_string(),
-            status: ContextRunStatus::Running,
-            step_id: Some("session-run".to_string()),
-            payload: serde_json::json!({
-                "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                "messageID": event.properties.get("messageID").cloned().unwrap_or(Value::Null),
-                "iteration": event.properties.get("iteration").cloned().unwrap_or(Value::Null),
-                "mode": event.properties.get("mode").cloned().unwrap_or(Value::Null),
-                "intent": event.properties.get("intent").cloned().unwrap_or(Value::Null),
-                "selectedToolCount": event.properties.get("selectedToolCount").cloned().unwrap_or(Value::Null),
-                "totalAvailableTools": event.properties.get("totalAvailableTools").cloned().unwrap_or(Value::Null),
-                "offeredTools": event.properties.get("offeredTools").cloned().unwrap_or(Value::Array(Vec::new())),
-                "hiddenByScope": event.properties.get("hiddenByScope").cloned().unwrap_or(Value::Array(Vec::new())),
-                "strictProjectionActive": event.properties.get("strictProjectionActive").cloned().unwrap_or(Value::Bool(false)),
-                "scopeAllowlist": event.properties.get("scopeAllowlist").cloned().unwrap_or(Value::Array(Vec::new())),
-                "tenantContext": event_tenant_context_value(event),
-                "why_next_step": "tool routing manifest recorded",
-                "step_status": "in_progress",
-            }),
-        }),
-        "policy.decision.recorded" => {
-            let record = event.properties.get("record")?;
-            let decision = record
-                .get("decision")
-                .and_then(Value::as_str)
-                .unwrap_or("unknown");
-            let tool = record
-                .get("tool")
-                .and_then(Value::as_str)
-                .unwrap_or("runtime");
-            Some(ContextRunEventAppendInput {
-                event_type: "policy_decision_recorded".to_string(),
-                status: if decision == "allow" {
-                    ContextRunStatus::Running
-                } else {
-                    ContextRunStatus::Blocked
-                },
-                step_id: Some("policy-decision".to_string()),
-                payload: serde_json::json!({
-                    "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                    "messageID": event.properties.get("messageID").cloned().unwrap_or(Value::Null),
-                    "runID": event.properties.get("runID").cloned().unwrap_or(Value::Null),
-                    "automationID": event.properties.get("automationID").cloned().unwrap_or(Value::Null),
-                    "tool": event.properties.get("tool").cloned().unwrap_or(Value::Null),
-                    "decisionID": event.properties.get("decisionID").cloned().unwrap_or(Value::Null),
-                    "record": record.clone(),
-                    "tenantContext": event_tenant_context_value(event),
-                    "why_next_step": format!("policy decision `{decision}` recorded for `{tool}`"),
-                    "step_status": if decision == "allow" {
-                        "in_progress"
-                    } else {
-                        "blocked"
-                    },
-                }),
-            })
-        }
-        "mutation.checkpoint.recorded" => {
-            let record = event.properties.get("record")?;
-            let tool = record
-                .get("tool")
-                .and_then(|value| value.as_str())
-                .unwrap_or("tool");
-            let outcome = record
-                .get("outcome")
-                .and_then(|value| value.as_str())
-                .unwrap_or("succeeded");
-            let changed_file_count = record
-                .get("changed_file_count")
-                .and_then(|value| value.as_u64())
-                .unwrap_or(0);
-            Some(ContextRunEventAppendInput {
-                event_type: "mutation_checkpoint_recorded".to_string(),
-                status: ContextRunStatus::Running,
-                step_id: Some("session-run".to_string()),
-                payload: serde_json::json!({
-                    "sessionID": event.properties.get("sessionID").cloned().unwrap_or(Value::Null),
-                    "messageID": event.properties.get("messageID").cloned().unwrap_or(Value::Null),
-                    "tool": event.properties.get("tool").cloned().unwrap_or(Value::Null),
-                    "record": record.clone(),
-                    "tenantContext": event_tenant_context_value(event),
-                    "why_next_step": format!(
-                        "mutation checkpoint for `{tool}` recorded with outcome `{outcome}` and {changed_file_count} changed files"
-                    ),
-                    "step_status": if matches!(outcome, "failed" | "blocked") {
-                        "blocked"
-                    } else {
-                        "in_progress"
-                    },
-                }),
-            })
-        }
-        _ => None,
-    }
-}
+#[path = "tasks_context_run.rs"]
+mod tasks_context_run;
+use tasks_context_run::*;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn runtime_event(event_type: &str, properties: Value, seq: u64) -> EngineEvent {
-        let envelope = tandem_types::RuntimeEventEnvelope::derive(seq, 1_000 + seq, &properties);
-        EngineEvent::new(event_type, properties).with_envelope(envelope)
-    }
-
-    #[tokio::test]
-    async fn routine_background_tasks_exit_without_runtime_when_startup_failed() {
-        let state = AppState::new_starting("routine-startup-guard-test".to_string(), true);
-        state.mark_failed("test_failed", "startup failed").await;
-
-        tokio::time::timeout(
-            Duration::from_millis(250),
-            run_routine_scheduler(state.clone()),
-        )
-        .await
-        .expect("scheduler should exit when startup has failed");
-
-        tokio::time::timeout(
-            Duration::from_millis(250),
-            run_routine_executor(state.clone()),
-        )
-        .await
-        .expect("executor should exit when startup has failed");
-    }
-
-    #[test]
-    fn session_context_run_event_input_maps_tool_effect_events() {
-        let input = session_context_run_event_input(&EngineEvent::new(
-            "tool.effect.recorded",
-            serde_json::json!({
-                "sessionID": "session-1",
-                "messageID": "message-1",
-                "tool": "write",
-                "record": {
-                    "session_id": "session-1",
-                    "message_id": "message-1",
-                    "tool": "write",
-                    "phase": "outcome",
-                    "status": "succeeded",
-                    "args_summary": {"path": "src/lib.rs"}
-                }
-            }),
-        ))
-        .expect("tool effect append input");
-
-        assert_eq!(input.event_type, "tool_effect_recorded");
-        assert_eq!(input.status, ContextRunStatus::Running);
-        assert_eq!(
-            input.payload.get("tool").and_then(Value::as_str),
-            Some("write")
-        );
-        assert_eq!(
-            input
-                .payload
-                .get("record")
-                .and_then(|value| value.get("status"))
-                .and_then(Value::as_str),
-            Some("succeeded")
-        );
-    }
-
-    #[test]
-    fn session_context_run_event_input_skips_running_tool_arg_deltas() {
-        let input = session_context_run_event_input(&EngineEvent::new(
-            "message.part.updated",
-            serde_json::json!({
-                "sessionID": "session-1",
-                "runID": "run-1",
-                "part": {
-                    "type": "tool",
-                    "tool": "write",
-                    "state": "running",
-                    "args": {"content": "{\"partial\":"}
-                },
-                "toolCallDelta": {
-                    "argsDelta": "large streamed write body chunk"
-                }
-            }),
-        ));
-
-        assert!(input.is_none());
-    }
-
-    #[test]
-    fn session_context_run_event_input_compacts_completed_tool_payloads() {
-        let oversized_content = "x".repeat(CONTEXT_TOOL_EVENT_STRING_LIMIT + 200);
-        let input = session_context_run_event_input(&EngineEvent::new(
-            "message.part.updated",
-            serde_json::json!({
-                "sessionID": "session-1",
-                "runID": "run-1",
-                "part": {
-                    "type": "tool",
-                    "tool": "write",
-                    "state": "completed",
-                    "args": {"content": oversized_content}
-                },
-                "toolCallDelta": Value::Null
-            }),
-        ))
-        .expect("completed tool update should still be journaled");
-
-        let compacted = input
-            .payload
-            .get("part")
-            .and_then(|part| part.get("args"))
-            .and_then(|args| args.get("content"))
-            .and_then(Value::as_str)
-            .expect("compacted content");
-        assert!(compacted.len() < CONTEXT_TOOL_EVENT_STRING_LIMIT + 100);
-        assert!(compacted.ends_with("...<truncated>"));
-    }
-
-    #[test]
-    fn session_context_run_event_input_maps_mutation_checkpoint_events() {
-        let input = session_context_run_event_input(&EngineEvent::new(
-            "mutation.checkpoint.recorded",
-            serde_json::json!({
-                "sessionID": "session-1",
-                "messageID": "message-1",
-                "tool": "write",
-                "record": {
-                    "session_id": "session-1",
-                    "message_id": "message-1",
-                    "tool": "write",
-                    "outcome": "succeeded",
-                    "file_count": 1,
-                    "changed_file_count": 1,
-                    "files": [{
-                        "path": "src/lib.rs",
-                        "resolved_path": "/workspace/src/lib.rs",
-                        "existed_before": false,
-                        "existed_after": true,
-                        "changed": true,
-                        "rollback_snapshot": {
-                            "status": "not_needed"
-                        }
-                    }]
-                }
-            }),
-        ))
-        .expect("mutation checkpoint append input");
-
-        assert_eq!(input.event_type, "mutation_checkpoint_recorded");
-        assert_eq!(
-            input.payload.get("tool").and_then(Value::as_str),
-            Some("write")
-        );
-        assert_eq!(
-            input
-                .payload
-                .get("record")
-                .and_then(|value| value.get("changed_file_count"))
-                .and_then(Value::as_u64),
-            Some(1)
-        );
-    }
-
-    #[tokio::test]
-    async fn runtime_event_log_enrichment_maps_session_only_events_to_run_and_tenant() {
-        let path = std::env::temp_dir().join(format!(
-            "runtime-events-persister-enrichment-{}.jsonl",
-            uuid::Uuid::new_v4()
-        ));
-        let tenant_context =
-            TenantContext::explicit_user_workspace("org-a", "workspace-a", None, "user-a");
-        let mut session = Session::new(Some("Session A".to_string()), Some(".".to_string()));
-        session.id = "session-a".to_string();
-        session.tenant_context = tenant_context.clone();
-        let mut context_cache = RuntimeEventLogContextCache::default();
-
-        let started = runtime_event(
-            "session.run.started",
-            serde_json::json!({
-                "sessionID": "session-a",
-                "runID": "run-a",
-                "tenantContext": tenant_context.clone(),
-            }),
-            1,
-        );
-        let started_row = RuntimeEventLogRow::from_engine_event(&started).expect("started row");
-        let started_row = enrich_runtime_event_log_row_from_session(
-            started_row,
-            Some(&session),
-            &mut context_cache,
-        );
-        append_runtime_event_log_row(&path, &started_row)
-            .await
-            .expect("append started");
-
-        let provider_iteration = runtime_event(
-            "provider.call.iteration.start",
-            serde_json::json!({
-                "sessionID": "session-a",
-                "provider": "openai",
-            }),
-            2,
-        );
-        let provider_row =
-            RuntimeEventLogRow::from_engine_event(&provider_iteration).expect("provider row");
-        assert_eq!(provider_row.run_id(), None);
-        assert_eq!(provider_row.tenant_context(), None);
-
-        let provider_row = enrich_runtime_event_log_row_from_session(
-            provider_row,
-            Some(&session),
-            &mut context_cache,
-        );
-        append_runtime_event_log_row(&path, &provider_row)
-            .await
-            .expect("append provider");
-
-        let rows = crate::runtime_event_log::query_runtime_event_log(
-            &path,
-            &session.tenant_context,
-            crate::runtime_event_log::RuntimeEventLogQuery {
-                run_id: "run-a",
-                after_seq: Some(1),
-                limit: None,
-            },
-        );
-
-        assert_eq!(rows.len(), 1);
-        assert_eq!(rows[0].seq(), 2);
-        assert_eq!(rows[0].run_id(), Some("run-a"));
-        assert_eq!(
-            rows[0]
-                .tenant_context()
-                .map(|tenant| tenant.org_id.as_str()),
-            Some("org-a")
-        );
-        assert_eq!(
-            rows[0].event.event_type.as_str(),
-            "provider.call.iteration.start"
-        );
-
-        let _ = tokio::fs::remove_file(path).await;
-    }
-}
+#[path = "tasks_core_tests.rs"]
+mod tests;
 
 pub async fn run_session_part_persister(state: AppState) {
     if !state.wait_until_ready_or_failed(120, 250).await {
@@ -940,6 +489,17 @@ async fn run_runtime_event_log_persister_with_registration_signal(
                     "runtime event log persister could not prune stale stateful waits"
                 );
             }
+        }
+
+        let reliability_path =
+            stateful_reliability_path_from_runtime_events_path(&state.runtime_events_path);
+        if let Err(error) =
+            prune_stateful_reliability_store(&reliability_path, retention_ms, now_ms()).await
+        {
+            tracing::warn!(
+                error = %error,
+                "runtime event log persister could not prune settled stateful reliability records"
+            );
         }
     }
