+++ crates/tandem-server/src/stateful_runtime/orchestration_store/runtime_records.rs
@@ -2,9 +2,14 @@ use anyhow::{bail, Context};
 use rusqlite::{params, OptionalExtension, TransactionBehavior};
 
 use super::OrchestrationStateStore;
-use crate::stateful_runtime::{
-    stateful_run_event_compacted_event_ids, StatefulRunEventRecord, StatefulRunSnapshotRecord,
+use crate::stateful_runtime::reliability::{
+    StatefulCompensationRecord, StatefulDeadLetterRecord, StatefulOutboxRecord,
+    StatefulReliabilityStoreFile, StatefulToolEffectRecord,
 };
+use crate::stateful_runtime::types::{
+    StatefulRunEventRecord, StatefulRunSnapshotRecord, StatefulWaitRecord,
+};
+use crate::stateful_runtime::{stateful_run_event_compacted_event_ids, StatefulRuntimeScope};
 
 impl OrchestrationStateStore {
     pub fn append_stateful_runtime_event(
@@ -164,6 +169,216 @@ impl OrchestrationStateStore {
                 .transpose()
         })
     }
+
+    pub fn load_stateful_runtime_waits(&self) -> anyhow::Result<Vec<StatefulWaitRecord>> {
+        self.with_connection(|connection| {
+            let mut statement = connection.prepare(
+                "SELECT wait_json FROM automation_waits
+                 ORDER BY updated_at_ms, wait_id, run_id",
+            )?;
+            let rows = statement.query_map([], |row| row.get::<_, String>(0))?;
+            rows.map(|row| serde_json::from_str(&row?).map_err(Into::into))
+                .collect()
+        })
+    }
+
+    pub fn replace_stateful_runtime_waits(
+        &self,
+        waits: &[StatefulWaitRecord],
+    ) -> anyhow::Result<()> {
+        self.with_connection(|connection| {
+            let transaction =
+                connection.transaction_with_behavior(TransactionBehavior::Immediate)?;
+            transaction.execute("DELETE FROM automation_waits", [])?;
+            for wait in waits {
+                insert_wait(&transaction, wait)?;
+            }
+            transaction.commit()?;
+            Ok(())
+        })
+    }
+
+    pub fn load_stateful_runtime_reliability(
+        &self,
+    ) -> anyhow::Result<StatefulReliabilityStoreFile> {
+        self.with_connection(|connection| {
+            Ok(StatefulReliabilityStoreFile {
+                schema_version: crate::stateful_runtime::STATEFUL_RUNTIME_SCHEMA_VERSION,
+                outbox: load_runtime_records(connection, "outbox_effects", "effect_json")?,
+                tool_effects: load_runtime_records(connection, "tool_effects", "effect_json")?,
+                dead_letters: load_runtime_records(connection, "dead_letters", "record_json")?,
+                compensations: load_runtime_records(connection, "compensations", "record_json")?,
+            })
+        })
+    }
+
+    pub fn replace_stateful_runtime_reliability(
+        &self,
+        reliability: &StatefulReliabilityStoreFile,
+    ) -> anyhow::Result<()> {
+        self.with_connection(|connection| {
+            let transaction =
+                connection.transaction_with_behavior(TransactionBehavior::Immediate)?;
+            for table in [
+                "outbox_effects",
+                "tool_effects",
+                "dead_letters",
+                "compensations",
+            ] {
+                transaction.execute(&format!("DELETE FROM {table}"), [])?;
+            }
+            for row in &reliability.outbox {
+                insert_reliability_record(
+                    &transaction,
+                    "outbox_effects",
+                    "effect_id",
+                    &row.outbox_id,
+                    row.run_id.as_deref(),
+                    &row.scope,
+                    &row.status,
+                    "effect_json",
+                    row.updated_at_ms,
+                    row,
+                )?;
+            }
+            for row in &reliability.tool_effects {
+                insert_reliability_record(
+                    &transaction,
+                    "tool_effects",
+                    "effect_id",
+                    &row.effect_id,
+                    row.run_id.as_deref(),
+                    &row.scope,
+                    &row.status,
+                    "effect_json",
+                    row.updated_at_ms,
+                    row,
+                )?;
+            }
+            for row in &reliability.dead_letters {
+                insert_reliability_record(
+                    &transaction,
+                    "dead_letters",
+                    "dead_letter_id",
+                    &row.dead_letter_id,
+                    row.run_id.as_deref(),
+                    &row.scope,
+                    &row.status,
+                    "record_json",
+                    row.updated_at_ms,
+                    row,
+                )?;
+            }
+            for row in &reliability.compensations {
+                insert_reliability_record(
+                    &transaction,
+                    "compensations",
+                    "compensation_id",
+                    &row.compensation_id,
+                    row.run_id.as_deref(),
+                    &row.scope,
+                    &row.status,
+                    "record_json",
+                    row.updated_at_ms,
+                    row,
+                )?;
+            }
+            transaction.commit()?;
+            Ok(())
+        })
+    }
+}
+
+fn insert_wait(
+    transaction: &rusqlite::Transaction<'_>,
+    wait: &StatefulWaitRecord,
+) -> anyhow::Result<()> {
+    transaction.execute(
+        "INSERT INTO automation_waits
+            (wait_id, goal_id, run_id, org_id, workspace_id, deployment_id,
+             status, wait_json, updated_at_ms)
+         VALUES (?1, NULL, ?2, ?3, ?4, ?5, ?6, ?7, ?8)
+         ON CONFLICT(wait_id, run_id, org_id, workspace_id, deployment_id) DO UPDATE SET
+            status = excluded.status, wait_json = excluded.wait_json,
+            updated_at_ms = excluded.updated_at_ms",
+        params![
+            wait.wait_id,
+            wait.run_id,
+            wait.scope.tenant_context.org_id,
+            wait.scope.tenant_context.workspace_id,
+            wait.scope
+                .tenant_context
+                .deployment_id
+                .as_deref()
+                .unwrap_or(""),
+            enum_name(&wait.status)?,
+            serde_json::to_string(wait)?,
+            wait.updated_at_ms,
+        ],
+    )?;
+    Ok(())
+}
+
+fn load_runtime_records<T>(
+    connection: &rusqlite::Connection,
+    table: &str,
+    json_column: &str,
+) -> anyhow::Result<Vec<T>>
+where
+    T: serde::de::DeserializeOwned,
+{
+    let mut statement = connection.prepare(&format!(
+        "SELECT {json_column} FROM {table} ORDER BY updated_at_ms, rowid"
+    ))?;
+    let rows = statement.query_map([], |row| row.get::<_, String>(0))?;
+    rows.map(|row| serde_json::from_str(&row?).map_err(Into::into))
+        .collect()
+}
+
+#[allow(clippy::too_many_arguments)]
+fn insert_reliability_record<T: serde::Serialize, S: serde::Serialize>(
+    transaction: &rusqlite::Transaction<'_>,
+    table: &str,
+    id_column: &str,
+    id: &str,
+    run_id: Option<&str>,
+    scope: &StatefulRuntimeScope,
+    status: &S,
+    json_column: &str,
+    updated_at_ms: u64,
+    record: &T,
+) -> anyhow::Result<()> {
+    let sql = format!(
+        "INSERT INTO {table}
+            ({id_column}, goal_id, run_id, status, {json_column}, updated_at_ms,
+             org_id, workspace_id, deployment_id)
+         VALUES (?1, NULL, ?2, ?3, ?4, ?5, ?6, ?7, ?8)
+         ON CONFLICT({id_column}) DO UPDATE SET status = excluded.status,
+             {json_column} = excluded.{json_column}, updated_at_ms = excluded.updated_at_ms,
+             org_id = excluded.org_id, workspace_id = excluded.workspace_id,
+             deployment_id = excluded.deployment_id"
+    );
+    transaction.execute(
+        &sql,
+        params![
+            id,
+            run_id,
+            enum_name(status)?,
+            serde_json::to_string(record)?,
+            updated_at_ms,
+            scope.tenant_context.org_id,
+            scope.tenant_context.workspace_id,
+            scope.tenant_context.deployment_id,
+        ],
+    )?;
+    Ok(())
+}
+
+fn enum_name<T: serde::Serialize>(value: &T) -> anyhow::Result<String> {
+    serde_json::to_value(value)?
+        .as_str()
+        .map(ToOwned::to_owned)
+        .ok_or_else(|| anyhow::anyhow!("serialized stateful status was not a string"))
 }
 
 fn insert_event(
